+++ packages/ethereum-indexer-streams/CHANGELOG.md
@@ -1,5 +1,13 @@
 # ethereum-indexer-server
 
+## 0.6.23
+
+### Patch Changes
+
+- Updated dependencies
+  - ethereum-indexer@0.6.18
+  - ethereum-indexer-utils@0.6.11
+
 ## 0.6.22
 
 ### Patch Changes